add item ownership verification for udpate and delete

+++ repository/repositories/item-repository.go
@@ -16,8 +16,8 @@ type ItemRepositoryInterface interface {
 	GetItemsList(userId int) ([]*models.Item, error)
 	CreateItem(srcItem *models.Item) (*models.Item, error)
 	GetItemById(id int, userId int) (*models.Item, error)
-	UpdateItem(id int, srcItem *models.Item) (*models.Item, error)
-	DeleteItem(id int) error
+	UpdateItem(id int, srcItem *models.Item, userId int) (*models.Item, error)
+	DeleteItem(id int, userId int) error
 }
 
 func GetItemRepository(log *log.Logger, connection database.Database) ItemRepositoryInterface {
@@ -80,10 +80,12 @@ func (r *ItemRepository) GetItemById(id int, userId int) (*models.Item, error) {
 	return &item, nil
 }
 
-func (r *ItemRepository) UpdateItem(id int, srcItem *models.Item) (*models.Item, error) {
+func (r *ItemRepository) UpdateItem(id int, srcItem *models.Item, userId int) (*models.Item, error) {
 	updateQuery :=
-		"UPDATE items SET user_id = $1, title = $2, initial_price = $3, description = $4 WHERE id = $5 RETURNING *"
-	row, err := r.db.Query(updateQuery, srcItem.UserId, srcItem.Title, srcItem.InitialPrice, srcItem.Description, id)
+		"UPDATE items SET user_id = $1, title = $2, initial_price = $3, description = $4 " +
+			"WHERE id = $5 AND user_id = $6 RETURNING *"
+	row, err := r.db.Query(updateQuery, userId, srcItem.Title, srcItem.InitialPrice,
+		srcItem.Description, id, userId)
 	if err != nil {
 		r.log.Error("failed to update item in db", err)
 		return nil, err
@@ -106,8 +108,8 @@ func (r *ItemRepository) UpdateItem(id int, srcItem *models.Item) (*models.Item,
 	return &updatedItem, nil
 }
 
-func (r *ItemRepository) DeleteItem(id int) error {
-	result, err := r.db.Exec("DELETE FROM items WHERE id = $1", id)
+func (r *ItemRepository) DeleteItem(id int, userId int) error {
+	result, err := r.db.Exec("DELETE FROM items WHERE id = $1 AND user_id = $2", id, userId)
 	if err != nil {
 		r.log.Error("failed to delete item from db", err)
 		return err
